pytorch_model/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==1.1.2
pandas==1.0.3
torchvision==0.7.0+cpu
nltk==3.4.5
spacy==2.3.2
torch==1.6.0+cpu
matplotlib==3.1.3
Pillow==8.0.1
torchtext==0.8.0
